requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
together==0.1.7
requests==2.31.0
python-dotenv==1.0.0
eventlet==0.33.3
SQLAlchemy==2.0.7
Werkzeug==2.2.3
Jinja2==3.1.2
Click==8.1.3
itsdangerous==2.1.2